refactor: move unofficial badge api to tmi module
