+++ rest-tmi/src/main/java/com/github/twitch4j/tmi/domain/BadgeSet.java
@@ -1,4 +1,4 @@
-package com.github.twitch4j.kraken.domain;
+package com.github.twitch4j.tmi.domain;
 
 import lombok.AccessLevel;
 import lombok.Data;
